Gmaps.java can now get bounds of map, references #4
Intial methods required to grab all 4 LatLng coordinates
of the static google map are implimented with the accuracy
that is required for this project.

Next step is to use these bounds to plot any LatLng
coordinates on the static map using the associations between
these points and the width and height in pixels of the image.
This will allow us to draw them at the proper pixel coordinates.