Google Maps Boundary and Gps buiding finished
GMaps.java when built to the GUI in its current state will
take a mapCenter, boundary top left, boundary bottom right, and
gps latitude and longitude location, and computer the visual
representation of the search area as well as the boundary that
the drone is searching in, as well as the location of the GPS
location, on the google static map.

More work is needed to make these components dynamically change
based on button action events.